Add MethodInvocation::metadata
