+++ src/Line/MethodInvocation/MethodInvocationInterface.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace webignition\BasilCompilationSource\Line\MethodInvocation;
 
 use webignition\BasilCompilationSource\Line\LineInterface;
+use webignition\BasilCompilationSource\Metadata\MetadataInterface;
 
 interface MethodInvocationInterface extends LineInterface
 {
@@ -16,4 +17,5 @@ interface MethodInvocationInterface extends LineInterface
     public function getArguments(): array;
 
     public function getArgumentFormat(): int;
+    public function getMetadata(): MetadataInterface;
 }
