+++ tests/Unit/Line/MethodInvocation/ObjectMethodInvocationTest.php
@@ -2,13 +2,15 @@
 
 declare(strict_types=1);
 
-namespace webignition\BasilCompilationSource\Tests\Unit\Line;
+namespace webignition\BasilCompilationSource\Tests\Unit\Line\MethodInvocation;
 
 use webignition\BasilCompilationSource\Line\LineTypes;
 use webignition\BasilCompilationSource\Line\MethodInvocation\ArgumentFormats;
-use webignition\BasilCompilationSource\Line\MethodInvocation\MethodInvocation;
+use webignition\BasilCompilationSource\Line\MethodInvocation\ObjectMethodInvocation;
+use webignition\BasilCompilationSource\Metadata\Metadata;
+use webignition\BasilCompilationSource\VariablePlaceholderCollection;
 
-class MethodInvocationTest extends \PHPUnit\Framework\TestCase
+class ObjectMethodInvocationTest extends \PHPUnit\Framework\TestCase
 {
     /**
      * @dataProvider createDataProvider
@@ -19,39 +21,45 @@ class MethodInvocationTest extends \PHPUnit\Framework\TestCase
      * @param string $expectedStringRepresentation
      */
     public function testCreate(
+        string $object,
         string $methodName,
         array $arguments,
         int $argumentFormat,
         string $expectedStringRepresentation
     ) {
-        $invocation = new MethodInvocation($methodName, $arguments, $argumentFormat);
+        $invocation = new ObjectMethodInvocation($object, $methodName, $arguments, $argumentFormat);
 
+        $this->assertSame($object, $invocation->getObject());
         $this->assertSame($methodName, $invocation->getMethodName());
         $this->assertSame($arguments, $invocation->getArguments());
         $this->assertSame($argumentFormat, $invocation->getArgumentFormat());
         $this->assertSame(LineTypes::METHOD_INVOCATION, $invocation->getType());
         $this->assertSame($expectedStringRepresentation, $invocation->getContent());
         $this->assertSame($expectedStringRepresentation, $invocation->__toString());
+        $this->assertEquals(new Metadata(), $invocation->getMetadata());
     }
 
     public function createDataProvider(): array
     {
         return [
             'no arguments' => [
+                'objectName' => 'object',
                 'methodName' => 'method',
                 'arguments' => [],
                 'argumentFormat' => ArgumentFormats::INLINE,
-                'expectedStringRepresentation' => 'method()'
+                'expectedStringRepresentation' => 'object->method()'
             ],
             'single argument' => [
+                'objectName' => 'object',
                 'methodName' => 'method',
                 'arguments' => [
                     1,
                 ],
                 'argumentFormat' => ArgumentFormats::INLINE,
-                'expectedStringRepresentation' => 'method(1)'
+                'expectedStringRepresentation' => 'object->method(1)'
             ],
             'multiple arguments, inline' => [
+                'objectName' => 'object',
                 'methodName' => 'method',
                 'arguments' => [
                     2,
@@ -59,9 +67,10 @@ class MethodInvocationTest extends \PHPUnit\Framework\TestCase
                     '"double-quoted value"'
                 ],
                 'argumentFormat' => ArgumentFormats::INLINE,
-                'expectedStringRepresentation' => 'method(2, \'single-quoted value\', "double-quoted value")'
+                'expectedStringRepresentation' => 'object->method(2, \'single-quoted value\', "double-quoted value")'
             ],
             'multiple arguments, stacked' => [
+                'objectName' => 'object',
                 'methodName' => 'method',
                 'arguments' => [
                     2,
@@ -69,8 +78,24 @@ class MethodInvocationTest extends \PHPUnit\Framework\TestCase
                     '"double-quoted value"'
                 ],
                 'argumentFormat' => ArgumentFormats::STACKED,
-                'expectedStringRepresentation' => 'method(2, \'single-quoted value\', "double-quoted value")'
+                'expectedStringRepresentation' => 'object->method(2, \'single-quoted value\', "double-quoted value")'
             ],
         ];
+    }
+
+    public function testWithMetadata()
+    {
+        $metadata = (new Metadata())
+            ->withVariableDependencies(VariablePlaceholderCollection::createCollection([
+                'variableDependencyOne',
+                'variableDependencyTwo',
+            ]));
+
+        $invocation = new ObjectMethodInvocation('object', 'methodName');
+        $this->assertEquals(new Metadata(), $invocation->getMetadata());
+
+        $mutatedInvocation = $invocation->withMetadata($metadata);
+        $this->assertSame($metadata, $mutatedInvocation->getMetadata());
+        $this->assertNotEquals($metadata, $invocation->getMetadata());
     }
 }
